Added menu bgm and door open close part 1
How long is a day and year on Venus?
Venus takes 224.7 Earth days to complete one orbit around the Sun. So a year on Venus is 0.62 as long as an Earth year. One interesting fact about Venus is that a day on Venus is actually longer than a year on Venus. A Venusian day lasts 243 days, and a year is 224 days.