added button press sfx
When was the Canadian flag made?
Out of three choices, the maple leaf design by George Stanley, based on the flag of the Royal Military College of Canada, was selected. The flag made its first official appearance on February 15, 1965; the date is now celebrated annually as National Flag of Canada Day.