Added Sound Effects Mixer
Is there a difference between a pig and a hog?
In the United States, the term "pig" refers to a younger domesticated swine weighing less than 120 pounds (50 kilograms), and the term "hog" refers to older swine weighing more than 120 lbs. In Great Britain all domesticated swine are referred to as pigs.