Menu button sounds. Imported main UI font
How many cells are there in an adult's body?
The mean volume of a mammal cell is estimated to be 4 billionths of a cubic centimeter. (To get a sense of that size, check out The Scale of the Universe.) Based on an adult man's typical volume, you might conclude that the human body contains 15 trillion cells.